% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Implements custom clustering algorithm}
\usage{
cluster(data, k, maxIteratoins = 100)
}
\arguments{
\item{data}{A data frame}

\item{k}{An integer indicating the number of clusters}

\item{maxIterations}{An integer indicating the number of iterations}
}
\value{
A list containing the cluster assignments
}
\description{
Implements custom clustering algorithm
}
