% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{kClusters}
\alias{kClusters}
\title{K-means clustering with random centroids}
\usage{
kClusters(data, k)
}
\arguments{
\item{data}{A matrix containing the 2 most principle components}

\item{k}{An integer indicating the number of clusters}
}
\value{
A list containing the clustering assignments and the centroids
}
\description{
K-means clustering with random centroids
}
