% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{principleComponents}
\alias{principleComponents}
\title{Converts data of many dimensions into 2-dimensional data containing the
2 most principle components}
\usage{
principleComponents(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A matrix containing the 2 most principle component
}
\description{
Converts data of many dimensions into 2-dimensional data containing the
2 most principle components
}
